Restructuring the code.
