+++ package.json
@@ -7,7 +7,7 @@
         "build": "webpack --config ./webpack.config.js --mode production",
         "wds": "webpack-dev-server",
         "start:dev": "npm run wds -- --config ./webpack.config.dev.js --mode development",
-        "start": "npm run wds -- --config ./webpack.config.js --mode production",
+        "start": "simplehttpserver dist",
         "test": "jest",
         "test:coverage": "npm run test -- --coverage"
     },
@@ -38,6 +38,7 @@
         "enzyme-adapter-react-16": "^1.1.1",
         "jest": "^22.4.3",
         "react-test-renderer": "^16.3.2",
+        "simplehttpserver": "^0.1.1",
         "style-loader": "^0.21.0",
         "webpack": "^4.5.0",
         "webpack-bundle-analyzer": "^2.11.1",
