+++ src/components/Movie/MovieBanner.js
