+++ src/components/Movie/MovieDetails.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 
-import Header from './Header/Header';
-import Footer from './Footer/Footer';
+import Header from '../Header/Header';
+import Footer from '../Footer/Footer';
 import SimilarMovies from './SimilarMovies';
 import MovieBanner from './MovieBanner';
 import './moviedetails.css';
