+++ src/components/Search/SearchFilter.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 
-import Button from './Common/Button';
+import Button from '../Common/Button';
 
 const SearchFilter = (props) => {
     return (
