+++ src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
 import ErrorBoundary from './components/Common/ErrorBoundary';
-import App from './components/App';
-import MovieDetails from './components/MovieDetails';
+import MainPage from './components/Search/MainPage';
+import MovieDetails from './components/Movie/MovieDetails';
 
 if (PRODUCTION === true)
     console.log("Running production build version", VERSION);
@@ -15,7 +15,7 @@ ReactDOM.render(
     <ErrorBoundary>
         <Router>
             <Switch>
-                <Route exact path="/" component={App} />
+                <Route exact path="/" component={MainPage} />
                 <Route path="/film/:id" component={MovieDetails} />
             </Switch>
         </Router>
